+++ soup/RenderTargetCanvas.hpp
@@ -10,6 +10,6 @@ namespace soup
 
 		RenderTargetCanvas(Canvas* canvas);
 
-		void drawRect(size_t x, size_t y, size_t width, size_t height, Rgb colour) final;
+		void drawRect(unsigned int x, unsigned int y, unsigned int width, unsigned int height, Rgb colour) final;
 	};
 }
